refactor : 상품, 게시물, 회원가입 이미지 API 변경
