+++ src/api/UploadImage.js
@@ -2,6 +2,7 @@ import axios from "axios";
 import { BASE_URL } from "../utils";
 import { imageInstance } from "./instance";
 
+// TODO: 이미지 API 3번째걸로 통일
 export const uploadImage = async (file) => {
   try {
     const formData = new FormData();
@@ -27,6 +28,7 @@ export const uploadImageAPI = async (file) => {
     const formData = new FormData();
     formData.append("image", file);
     const res = await imageInstance.post("/image/uploadfile", formData);
+    console.log(res);
     return res.data.filename;
   } catch (error) {
     throw error;
